Restoring original recommends on iceweasel, now that it has improved a lot.
In squeeze lxde recommended iceweasel. During the wheezy release
cycle, the firefox development got stalled at some point and
chromium was a quicker, more responsive browser at that time.
So lxde was updated to recommend chromium instead.

However, the mozilla project increased pace and improved a lot
for firefox version 4 and newer, bringing firefox back on pair
with chromium. Therefore this change does not reopen #570176.

Having chromium and iceweasel more or less identical we think it
is better to by default pull in iceweasel since it's the default
on all other desktop environments.

In future (wheezy+1) we might reconsider and recommend a more
leightweight browser by default. Currently iceweasel (or chromium)
is the better recommendation for users, providing a solid 'general
purpose' browser rather than being an exotic niche.