Switching from gdm3 to lightdm.
gdm3 had recently a depends added on some more gnome packages,
so effectively having a recommends on gdm3 means to pull in half
of gnome, which we obviously don't want.

The only reasonable alternative is to use lightdm for wheezy,
as lxdm is not ready yet. tasksel will also use lightdm for wheezy
in the task-lxde package.